+++ packages/wonder-blocks-tokens/src/theme/semantic/semantic-color.ts
@@ -43,11 +43,6 @@ const core = {
             default: color.fadedOffBlack16,
             strong: color.fadedOffBlack32,
         },
-        inverse: {
-            subtle: color.fadedOffBlack16,
-            default: color.fadedOffBlack8,
-            strong: color.white,
-        },
         knockout: {
             default: color.white,
         },
@@ -127,11 +122,6 @@ const core = {
             default: color.fadedOffBlack32,
             strong: color.fadedOffBlack50,
         },
-        inverse: {
-            subtle: color.fadedOffBlack32,
-            default: color.offWhite,
-            strong: color.white,
-        },
         knockout: {
             default: color.white,
         },
@@ -154,14 +144,6 @@ const core = {
             },
         },
     },
-};
-
-const surface = {
-    primary: color.white,
-    secondary: color.offWhite,
-    emphasis: color.fadedBlue16,
-    inverse: color.darkBlue,
-    overlay: color.offBlack64,
 };
 
 const sharedFeedbackStrongTokens = {
@@ -730,11 +712,6 @@ export const semanticColor = {
             foreground: core.foreground.neutral.strong,
         },
     },
-    /**
-     * For background colors. Overlays act as supplementary surfaces that mask
-     * areas of the UI.
-     */
-    surface,
 
     focus: {
         outer: color.blue,