+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-popover
 
+## 6.1.32
+
+### Patch Changes
+
+- Updated dependencies [94351f5]
+    - @khanacademy/wonder-blocks-icon-button@10.5.0
+    - @khanacademy/wonder-blocks-modal@8.4.2
+    - @khanacademy/wonder-blocks-tooltip@4.1.46
+
+## 6.1.31
+
+### Patch Changes
+
+- Updated dependencies [ede6085]
+- Updated dependencies [a70f274]
+    - @khanacademy/wonder-blocks-tokens@13.0.0
+    - @khanacademy/wonder-blocks-icon-button@10.4.2
+    - @khanacademy/wonder-blocks-modal@8.4.1
+    - @khanacademy/wonder-blocks-styles@0.2.31
+    - @khanacademy/wonder-blocks-tooltip@4.1.45
+    - @khanacademy/wonder-blocks-typography@4.2.21
+
 ## 6.1.30
 
 ### Patch Changes