+++ packages/wonder-blocks-tabs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @khanacademy/wonder-blocks-tabs
 
+## 0.3.29
+
+### Patch Changes
+
+- Updated dependencies [ede6085]
+- Updated dependencies [a70f274]
+    - @khanacademy/wonder-blocks-tokens@13.0.0
+    - @khanacademy/wonder-blocks-typography@4.2.21
+
 ## 0.3.28
 
 ### Patch Changes